% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{has_internet}
\alias{has_internet}
\title{Check for internet connection}
\usage{
has_internet(url = "https://www.google.com")
}
\arguments{
\item{url}{character.}
}
\value{
logical
}
\description{
A basic check for internet connectivity.
}
\examples{
has_internet()
}
