% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_alpha.R
\name{ma_article}
\alias{ma_article}
\title{Read Memory Alpha article}
\usage{
ma_article(url, browse = FALSE, data_format = c("xml_nodelist",
  "data_frame"))
}
\arguments{
\item{url}{character, article URL. Expects package-style short URL. See examples.}

\item{browse}{logical, also open \code{url} in browser.}

\item{data_format}{\code{"xml"} or \code{"df"}. A list is returned containing the character string title, and a data object as an XML nodelist or a data frame.}
}
\value{
a character string of article HTML.
}
\description{
Read Memory Alpha article
}
\examples{
\dontrun{ma_article("Worf")}
}
