% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_beta.R
\name{mb_search}
\alias{mb_search}
\title{Memory Beta site search}
\usage{
mb_search(text, browse = FALSE)
}
\arguments{
\item{text}{character, search query.}

\item{browse}{logical, open search results page in browser.}
}
\value{
a data frame
}
\description{
Perform a Memory Beta site search.
}
\details{
This function returns a data frame containing the title, truncated text preview, and relative URL for the first page of search results.
It does not recursively collate search results through subsequent pages of results.
There could be an unexpectedly high number of pages of results depending on the search query.
Since the general nature of this search feature seems relatively casual anyway, it aims only to provide a first page preview.
}
\examples{
\donttest{mb_search("Worf")}
}
