% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_alpha.R
\name{mem_alpha}
\alias{mem_alpha}
\title{Memory Alpha API}
\usage{
mem_alpha(id, by = NULL)
}
\arguments{
\item{id}{character, category ID. See details.}

\item{by}{character, describes how terminal pages are grouped by subcategories.}
}
\value{
a data frame.
}
\description{
Access data from Memory Alpha (\url{http://memory-alpha.wikia.com}).
}
\details{
Access data from Memory Alpha web pages using a category ID.
Subcategories are accessed by appending the specific subcategory's ID to \code{id} with a \code{:} separator.
Available top-level category IDs and associated subcategory options for \code{by} can be found by calling \code{mem_alpha} with no arguments.
See examples.

The content returned is always a data frame (except when calling with no arguments).
When the \code{id} request refers to a category, the data frame contains the names and associated URLs of entries in the category.
Note that the URLs are relative. To visit a URL directly, append \code{http://memory-alpha.wikia.com/}.

When the \code{id} request refers to a terminal content page rather than a category page,
the data frame contains the main text and other data associated with the entry.
}
\examples{
# show available top-level category IDs and associated subcategory options
mem_alpha()

# people by species category
mem_alpha("people", by = "species")

# subcategories and main pages under Klingon people-by-species subcategory
mem_alpha("people:Klingons", by = "species")
}
