% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrek.R
\docType{package}
\name{rtrek}
\alias{rtrek-package}
\alias{rtrek}
\title{rtrek: Data Analysis Relating to Star Trek}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides datasets related to the Star Trek fictional universe and functions for working with the data. The package also provides access to real world datasets based on the televised series and other related licensed media productions. It interfaces with the Star Trek API (STAPI) (\url{http://stapi.co/}), Memory Alpha (\url{https://memory-alpha.fandom.com/wiki/Portal:Main}), and Memory Beta (\url{https://memory-beta.fandom.com/wiki/Main_Page}) to retrieve data, metadata and other information relating to Star Trek. It also contains several local datasets covering a variety of topics. The package also provides functions for working with data from other Star Trek-related R data packages containing larger datasets not stored in 'rtrek'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonawicz/rtrek}
  \item Report bugs at \url{https://github.com/leonawicz/rtrek/issues}
}

}
\author{
\strong{Maintainer}: Matthew Leonawicz \email{rpkgs@pm.me} (\href{https://orcid.org/0000-0001-9452-2771}{ORCID})

}
