% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\name{st_book_series}
\alias{st_book_series}
\title{Go to Wikipedia entry for a specific book series}
\usage{
st_book_series(id)
}
\arguments{
\item{id}{character, abbreviation for a series. See details.}
}
\value{
opens a browser tab, nothing is returned unless \code{id} is not provided, in which case a data frame is returned.
}
\description{
This function opens a browser tab to the main Wikipedia entry for all Star Trek novels automatically scrolled to the section pertaining to \code{id}.
To see the available IDs, call \code{st_book_series} with no arguments.
}
\examples{
st_book_series()
st_book_series("DS9")
}
\seealso{
\code{\link{stBooks}}
}
