% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\name{st_books_wiki}
\alias{st_books_wiki}
\title{Go to Wikipedia entry for a specific book series}
\usage{
st_books_wiki()
}
\value{
opens a browser tab, nothing is returned.
}
\description{
This function opens a browser tab to the main Wikipedia entry for all Star Trek novels.
For a more complete set of Star Trek series, miniseries and anthology names and acronyms, see the \code{\link{stSeries}} and \code{\link{stBooks}} datasets.
}
\examples{
\dontrun{st_books_wiki()}
}
\seealso{
\code{\link{stBooks}} \code{\link{stSeries}}
}
