% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{st_tiles_data}
\alias{st_tiles_data}
\title{Ancillary location data for map tiles}
\usage{
st_tiles_data(id)
}
\arguments{
\item{id}{character, name of a map tile set.}
}
\value{
a data frame
}
\description{
Obtain a table of ancillary data associated with various locations of interest, given a specific map tile set ID.
}
\details{
This function returns a small example data frame of location-specific data along with grid cell coordinates that are specific to the requested map tile set ID.
}
\examples{
st_tiles_data("galaxy2")
}
\seealso{
\code{\link{stTiles}}, \code{\link{st_tiles}}
}
