% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{stapi_options}
\alias{stapi_options}
\title{List available STAPI entities}
\usage{
stapi_options()
}
\value{
a character vector.
}
\description{
Currently, this function simply lists the available entities for the Star Trek API (STAPI). See \code{\link{stapi}}.
}
\examples{
stapi_options()
}
\seealso{
\code{\link{stapi}}
}
