% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stBooksWiki}
\alias{stBooksWiki}
\title{Star Trek novel metadata from Wikipedia.}
\format{A data frame}
\usage{
stBooksWiki
}
\description{
A data frame with 719 rows and 6 columns containing metadata on Star Trek novels and other books taken from the primary Wikipedia page: \url{https://en.wikipedia.org/wiki/List_of_Star_Trek_novels}.
The data frame contains most of the novels but is not comprehensive, containing only the most easily scraped HTML table data, and may be out of date temporarily whenever new novels are published.
}
\details{
There is considerable overlap in titles between \code{stBooksWiki} and \code{stBooks}, but also unique entries and they offer some different fields.
}
\seealso{
\code{\link{st_books_wiki}}, \code{\link{stBooks}}
}
\keyword{datasets}
