% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stBooksWP}
\alias{stBooksWP}
\title{Star Trek novel metadata from Wikipedia.}
\format{A data frame}
\usage{
stBooksWP
}
\description{
A data frame with 715 rows and 6 columns containing metadata on Star Trek novels and other books taken from the primary Wikipedia page: https://en.wikipedia.org/wiki/List_of_Star_Trek_novels.
The data frame contains most of the novels but is not comprehensive, containing only the most easily scraped HTML table data, and may be out of date temporarily whenever new novels are published.
}
\details{
There is considerable overlap in titles between \code{stBooksWP} and \code{stBooks}, but also a considerable number of unique entries.
For example, the old novelizations by James Blish of the Original Series episodes are not found in \code{stBooks}.
}
\seealso{
\code{\link{st_book_series}}, \code{\link{stBooks}}
}
\keyword{datasets}
