% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stLogos}
\alias{stLogos}
\title{Star Trek logos metadata.}
\format{
A data frame
}
\usage{
stLogos
}
\description{
A data frame with 236 rows and 3 columns containing Star Trek logo metadata: category, description and URL. Logo artwork credited to Kris Trigwell.
The logo images are served by st-minutiae.com for personal and fair use.
}
\seealso{
\code{\link{st_logo}}
}
\keyword{datasets}
