% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stSeries}
\alias{stSeries}
\title{Star Trek series.}
\format{
A data frame
}
\usage{
stSeries
}
\description{
A data frame with 35 rows and 3 columns containing names and abbreviations of Star Trek series and anthologies.
There are so many because the table pertains to written works, which is inclusive of the more limited televised series.
}
\details{
Some entries listed as series can be interpreted as miniseries, but that distinction is not made here.
The official line between the two is not always clear and can also change as more novels are released.

Anthologies are listed as such, rather than as series. Reference manuals have a distinct entry.
The Miscellaneous category can be considered synonymous with All-Series/Crossover,
abbreviated elsewhere as simply \code{ST} for Star Trek in general, rather than as \code{MISC}.
}
\seealso{
\code{\link{st_books_wiki}}
}
\keyword{datasets}
