% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_rtrek}
\alias{theme_rtrek}
\alias{theme_rtrek_dark}
\title{ggplot2 themes}
\usage{
theme_rtrek(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_rtrek_dark(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size.}

\item{base_family}{base font family.}

\item{base_line_size}{base size for line elements.}

\item{base_rect_size}{base size for rect elements.}
}
\description{
A collection of ggplot2 themes.
}
