% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tlFootnotes}
\alias{tlFootnotes}
\title{Star Trek timeline footnotes.}
\format{A data frame}
\usage{
tlFootnotes
}
\description{
A data frame with 605 rows and 3 columns containing footnotes associated by ID with various entries in package timeline datasets, \code{tlBooks} and \code{tlEvents}.
}
\seealso{
\code{\link{tlBooks}} \code{\link{tlEvents}}
}
\keyword{datasets}
