% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrek.R
\docType{package}
\name{rtrek}
\alias{rtrek}
\alias{rtrek-package}
\title{rtrek: Star Trek datasets and related R functions.}
\description{
The rtrek package contains a collection of Star Trek-themed datasets and R functions to assist with their use.
}
