% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrek.R
\docType{package}
\name{rtrek}
\alias{rtrek}
\alias{rtrek-package}
\title{rtrek: Star Trek datasets and related R functions.}
\description{
The rtrek package contains a collection of Star Trek-themed datasets and R functions to assist with their use.
The package interfaces with Wikipedia, Memory Alpha and Memory Beta to retrieve data, metadata and other information relating to the Star Trek fictional universe.
It also contains local datasets covering a variety of topics such as Star Trek universe species data, geopolitical data, and datasets resulting from text mining analyses of Star Trek novels.
}
