% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stBooks}
\alias{stBooks}
\title{Star Trek novel metadata.}
\format{A data frame}
\usage{
stBooks
}
\description{
A data frame of with 715 row and 6 columns containing metadata on Star Trek novels and other books taken from the primary Wikipedia page: https://en.wikipedia.org/wiki/List_of_Star_Trek_novels.
The data frame contains most of the novels but is not comprehensive, containing only the most easily scraped HTML table data, and may be out of date temporarily whenever new novels are published.
}
\seealso{
\code{\link{st_book_series}}
}
\keyword{datasets}
