% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stGeo}
\alias{stGeo}
\title{Raster grid location data for stellar cartographic map tile set 1.}
\format{A data frame}
\usage{
stGeo
}
\description{
A data frame of with 7 rows and 6 columns.
}
\keyword{datasets}
