% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrek.R
\docType{data}
\name{STgeo}
\alias{STgeo}
\title{Raster grid location data for stellar cartographic map tile set 1.}
\format{A data frame}
\usage{
STgeo
}
\description{
A data frame of with 7 rows and 6 columns.
}
\keyword{datasets}
