% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stSpecies}
\alias{stSpecies}
\title{Species names and avatars, linked from Memory Alpha.}
\format{A data frame}
\usage{
stSpecies
}
\description{
A data frame of with 7 rows and 2 columns.
}
\keyword{datasets}
