% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrek.R
\docType{data}
\name{STspecies}
\alias{STspecies}
\title{Species name and avatars, linked from Memory Alpha.}
\format{A data frame}
\usage{
STspecies
}
\description{
A data frame of with 7 rows and 2 columns.
}
\keyword{datasets}
